% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calSim}
\alias{calSim}
\title{Function to calculate the similarity matrix based on the
cluster membership indicator of each iteration.}
\usage{
calSim(mat)
}
\arguments{
\item{mat}{Matrix of cluster membership indicator from all iterations}
}
\description{
Function to calculate the similarity matrix based on the
cluster membership indicator of each iteration.
}
\examples{
n = 90 ##number of subjects
iters = 200 ##number of iterations
## matrix of cluster membership indicators
## perfect clustering with three clusters
mat = matrix(rep(1:3,each=n/3),nrow=n,ncol=iters)
sim = calSim(t(mat))
}
